backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg[binary]==3.1.18
pydantic==1.10.13
python-dotenv==1.0.1
alembic==1.13.1
